+++ socket-server/lib/join-room-handler.js
@@ -1,0 +1,3 @@
+module.exports = async function joinRoomHandler(io, socket, roomName) {
+  socket.join(roomName);
+};