+++ socket-server/lib/leave-room-handler.js
@@ -1,0 +1,3 @@
+module.exports = async function leaveRoomHandler(io, socket, roomName) {
+  socket.leave(roomName);
+};