+++ src/app/app/(secondary)/event/(with-back-button)/create/from_template/page.tsx
@@ -1,19 +1,36 @@
+import { RoundButton } from '@/components/ui/round-button';
 import db from '@/dbconfig';
 import { TemplateLinkList } from '@/features/events/components/template-list';
 import { eventService } from '@/features/events/services/event-service';
 import { SearchProvider } from '@/providers/search-provider';
 import { tablenames } from '@/tablenames';
 import { loadSession } from '@/util/load-session';
+import { ArrowLeft, X } from 'lucide-react';
+import Link from 'next/link';
 
 export default async function EventTemplatesPage({ searchParams }) {
   const { q } = await searchParams;
   const session = await loadSession();
   const templates = await eventService.repo.findTemplatesByAuthorId(session.user.id, null, db);
   return (
-    <div className='flex flex-col gap-2 w-full py-2 px-default h-full'>
-      <SearchProvider>
-        <TemplateLinkList templates={templates} />
-      </SearchProvider>
-    </div>
+    <>
+      <main className='flex flex-col gap-2 w-full py-2 px-default flex-1 overflow-y-scroll'>
+        <SearchProvider>
+          <TemplateLinkList templates={templates} />
+        </SearchProvider>
+      </main>
+      <div className='w-full flex p-4 items-center justify-center gap-4'>
+        <Link href='/app/event/create'>
+          <RoundButton>
+            <ArrowLeft />
+          </RoundButton>
+        </Link>
+        <Link href={'/app/feed'}>
+          <RoundButton>
+            <X />
+          </RoundButton>
+        </Link>
+      </div>
+    </>
   );
 }